# lxhcool-backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/public/posts/{slug}": {
      "get": {
        "operationId": "PublicPostsController_getBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/PostResponseDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Public Posts"
        ]
      }
    },
    "/admin/posts": {
      "get": {
        "operationId": "AdminPostsController_list",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PostResponseDto"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Admin Posts"
        ]
      },
      "post": {
        "operationId": "AdminPostsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/PostResponseDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Admin Posts"
        ]
      }
    },
    "/admin/posts/{id}": {
      "get": {
        "operationId": "AdminPostsController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/PostResponseDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Admin Posts"
        ]
      },
      "patch": {
        "operationId": "AdminPostsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/PostResponseDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Admin Posts"
        ]
      },
      "delete": {
        "operationId": "AdminPostsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/PostResponseDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Admin Posts"
        ]
      }
    },
...
    "/admin/music/netease/metadata": {
      "get": {
        "operationId": "MusicController_getNeteaseMetadata",
        "parameters": [
          {
            "name": "url",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/NeteaseMetadataResponseDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Admin Music"
        ]
      }
    },
    "/admin/music/netease/playlist": {
      "get": {
        "operationId": "MusicController_getNeteasePlaylist",
        "parameters": [
          {
            "name": "url",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/NeteasePlaylistResponseDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Admin Music"
        ]
      }
    },
    "/public/music/netease/metadata": {
      "get": {
        "operationId": "PublicMusicController_getNeteaseMetadata",
        "parameters": [
          {
            "name": "url",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/NeteaseMetadataResponseDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Public Music"
        ]
      }
    },
    "/public/music/netease/playlist": {
      "get": {
        "operationId": "PublicMusicController_getNeteasePlaylist",
        "parameters": [
          {
            "name": "url",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/NeteasePlaylistResponseDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Public Music"
        ]
      }
    }
  },
  "info": {
    "title": "lxhcool API",
    "description": "API documentation for the lxhcool frontend and admin applications.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "NeteaseMetadataResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "artist": {
            "type": "string"
          },
          "cover": {
            "type": "string"
          },
          "duration": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "title",
          "artist",
          "cover",
          "duration"
        ]
      },
      "NeteasePlaylistTrackResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "artist": {
            "type": "string"
          },
          "cover": {
            "type": "string"
          },
          "duration": {
            "type": "number"
          },
          "externalUrl": {
            "type": "string"
          },
          "embedUrl": {
            "type": "string"
          },
          "audioUrl": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "artist",
          "cover",
          "duration",
          "externalUrl",
          "embedUrl",
          "audioUrl"
        ]
      },
      "NeteasePlaylistResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "cover": {
            "type": "string"
          },
          "trackCount": {
            "type": "number"
          },
          "tracks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NeteasePlaylistTrackResponseDto"
            }
          }
        },
        "required": [
          "id",
          "title",
          "cover",
          "trackCount",
          "tracks"
        ]
      },
...
      "PostResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "ARTICLE",
              "MOMENT"
            ]
          },
          "title": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "nullable": true
          },
          "excerpt": {
            "type": "string",
            "nullable": true
          },
          "content": {
            "type": "string"
          },
          "media": {
            "type": "object",
            "additionalProperties": true
          },
          "coverImage": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PUBLISHED",
              "HIDDEN"
            ]
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "seoTitle": {
            "type": "string",
            "nullable": true
          },
          "seoDescription": {
            "type": "string",
            "nullable": true
          },
          "ogImage": {
            "type": "string",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "type",
          "title",
          "slug",
          "category",
          "excerpt",
          "content",
          "media",
          "coverImage",
          "status",
          "publishedAt",
          "seoTitle",
          "seoDescription",
          "ogImage",
          "tags",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreatePostDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "ARTICLE",
              "MOMENT"
            ]
          },
          "title": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "excerpt": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "media": {
            "type": "object",
            "additionalProperties": true
          },
          "coverImage": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PUBLISHED",
              "HIDDEN"
            ]
          },
          "seoTitle": {
            "type": "string"
          },
          "seoDescription": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "title",
          "slug",
          "content"
        ]
      },
      "UpdatePostDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "ARTICLE",
              "MOMENT"
            ]
          },
          "title": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "excerpt": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "media": {
            "type": "object",
            "additionalProperties": true
          },
          "coverImage": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PUBLISHED",
              "HIDDEN"
            ]
          },
          "seoTitle": {
            "type": "string"
          },
          "seoDescription": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}